+++ tests/logger.test.js
@@ -23,7 +23,7 @@ describe("putLogEvent",  () => {
     });
 
     afterEach(() => {
-        fs.rmdirSync(baseRepoPath, {recursive: true});
+        fs.rmSync(baseRepoPath, {recursive: true});
     });
 
     test("No User", () => {